requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
langchain>=0.0.300
langchain-core>=0.1.21,<0.2
openai>=1.0.0
boto3>=1.28.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl==3.1.2
python-pptx>=0.6.21
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
typing-extensions>=4.6.0
pydantic<2.0.0
SQLAlchemy==2.0.38
unidecode>=1.3.6
chardet>=5.0.0
